requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
Flask
Flask-SQLAlchemy==3.0.5
frozenlist
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.1
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
requests==2.31.0
SQLAlchemy==2.0.19
twilio==8.5.0
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug
yarl
python-dotenv
flask_cors